[SPARK-17273][SQL] Move expression optimizer rules into a separate file
